+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v3.17.2 - 2018-04-09
 --------------------
 
 + Fixed issue preventing lesson video and audio embeds from being *removed* when using the course builder settings editor
++ Fixed issue causing question images to lose the image source
 + Updated student management table for courses and memberships to show the name (and a link to the user profile) of the site user who manually enrolled the student.
 + Add "All Time" reporting to various reporting filters
 + Added API for builder fields to enable multiple select fields